8303362: Serial: Move CardTableRS to serial directory
