+++ src/hotspot/share/gc/shared/generationSpec.cpp
@@ -23,7 +23,7 @@
  */
 
 #include "precompiled.hpp"
-#include "gc/shared/cardTableRS.hpp"
+#include "gc/serial/cardTableRS.hpp"
 #include "gc/shared/generationSpec.hpp"
 #include "runtime/java.hpp"
 #include "utilities/macros.hpp"
